Rename key-value expression classes
