+++ tests/Alchemy/Tests/Phrasea/SearchEngine/AST/EqualExpressionTest.php
@@ -2,7 +2,7 @@
 
 namespace Alchemy\Tests\Phrasea\SearchEngine\AST;
 
-use Alchemy\Phrasea\SearchEngine\Elastic\AST\FieldEqualsExpression;
+use Alchemy\Phrasea\SearchEngine\Elastic\AST\KeyValue\EqualExpression;
 use Alchemy\Phrasea\SearchEngine\Elastic\AST\KeyValue\Key;
 use Alchemy\Phrasea\SearchEngine\Elastic\Search\QueryContext;
 use Alchemy\Phrasea\SearchEngine\Elastic\Structure\Field as StructureField;
@@ -16,10 +16,10 @@ class FieldEqualsExpressionTest extends \PHPUnit_Framework_TestCase
 {
     public function testSerialization()
     {
-        $this->assertTrue(method_exists(FieldEqualsExpression::class, '__toString'), 'Class does not have method __toString');
+        $this->assertTrue(method_exists(EqualExpression::class, '__toString'), 'Class does not have method __toString');
         $key = $this->prophesize(Key::class);
         $key->__toString()->willReturn('foo');
-        $node = new FieldEqualsExpression($key->reveal(), 'bar');
+        $node = new EqualExpression($key->reveal(), 'bar');
         $this->assertEquals('(<foo> == <value:"bar">)', (string) $node);
     }
 
@@ -35,7 +35,7 @@ class FieldEqualsExpressionTest extends \PHPUnit_Framework_TestCase
         $key->getIndexField($query_context, true)->willReturn($index_field);
         // TODO Test keys implementing QueryPostProcessor
 
-        $node = new FieldEqualsExpression($key->reveal(), 'bar');
+        $node = new EqualExpression($key->reveal(), 'bar');
         $query = $node->buildQuery($query_context);
 
         $this->assertEquals(json_decode($expected_json, true), $query);
